+++ __tests__/supergraph/errors/INTERFACE_FIELD_NO_IMPLEM.spec.ts
@@ -2,7 +2,7 @@ import { expect, test } from 'vitest';
 import { graphql, testVersions } from '../../shared/testkit.js';
 
 testVersions((api, version) => {
-  test.skipIf(api.library === 'guild')('INTERFACE_FIELD_NO_IMPLEM', () => {
+  test('INTERFACE_FIELD_NO_IMPLEM (entity)', () => {
     expect(
       api.composeServices([
         {
@@ -49,7 +49,7 @@ testVersions((api, version) => {
         errors: expect.arrayContaining([
           expect.objectContaining({
             message: expect.stringContaining(
-              `Interface field "User.email" is declared in subgraph "users" but type "Author", which implements "User" only in subgraph "feed" does not have field "email".`,
+              `Interface field "User.email" is declared in subgraph "users" but type "Author", which implements "User" ${api.library === 'apollo' ? 'only ' : ''}in subgraph "feed" does not have field "email".`,
             ),
             extensions: expect.objectContaining({
               code: 'INTERFACE_FIELD_NO_IMPLEM',
@@ -58,7 +58,59 @@ testVersions((api, version) => {
         ]),
       }),
     );
+  });
 
-    // KNOW: check all interface fields are implemented when one subgraph introduces a new field to the interface
+  test('INTERFACE_FIELD_NO_IMPLEM (data)', () => {
+    expect(
+      api.composeServices([
+        {
+          name: 'foo',
+          typeDefs: graphql`
+            type Query {
+              foo: Foo
+            }
+
+            type Foo implements Person {
+              name: String
+              age: Int
+            }
+
+            interface Person {
+              name: String
+              age: Int
+            }
+          `,
+        },
+        {
+          name: 'bar',
+          typeDefs: graphql`
+            type Query {
+              bar: Bar
+            }
+
+            type Bar implements Person {
+              name: String
+            }
+
+            interface Person {
+              name: String
+            }
+          `,
+        },
+      ]),
+    ).toEqual(
+      expect.objectContaining({
+        errors: expect.arrayContaining([
+          expect.objectContaining({
+            message: expect.stringContaining(
+              `Interface field "Person.age" is declared in subgraph "foo" but type "Bar", which implements "Person" ${api.library === 'apollo' ? 'only ' : ''}in subgraph "bar" does not have field "age".`,
+            ),
+            extensions: expect.objectContaining({
+              code: 'INTERFACE_FIELD_NO_IMPLEM',
+            }),
+          }),
+        ]),
+      }),
+    );
   });
 });